Preserve and restore bar visibility states during project reset in ApplicationLogic. This change ensures that the visibility settings for heatmap and funscript timeline are maintained across resets, improving user experience.
Preserve and restore bar visibility states during project reset in ApplicationLogic. This change ensures that the visibility settings for heatmap and funscript timeline are maintained across resets, improving user experience.
